Component manager discovery
This commit removes the manual component manager registry system and replaces it with a system that looks up component managers through the container. This means that component managers are looked up lazily and only instantiated the first time a component with that manager is seen.

Additionally, this commit introduces instance initializers, and makes some tweaks to rationalize application initialization and boot. The application now implements the RegistryWriter interface, but only allows writing to the registry before initialization has completed.

Lastly, there are some changes to better support the template-only component case. In particular, component definitions are now cached internally based on the template specifier, not the component specifier.