Instance initializers -> initializers
Per @dgeb’s feedback (which I agree with wholeheartedly), I’ve changed the “initializer-that-receives-the-registry” to be called “initializer” instead of “instance initializer”. If the registry is write-only and doesn’t depend on application state, we can easily move this to something that is shared across all instances in the future.

I also eliminated the Application’s implementation of the RegistryWriter interface. Instead, we pass the `ApplicationRegistry` directly. One tradeoff here is that the app can close over the registry and continue adding entries even after the app has finished initializing. In the future, we may want to explore having the facility to freeze a registry.