Fix error being triggered by rollup during rebuild.
For the initial build, when rollup sees modules that it cannot
resolve it treates them as externals (and therefore they are runtime
required), however when rollup is seeded by a bundle (e.g. all rebuilds)
it errors when it hits externals that were not specified.

This is likely a bug in rollup...