Component `args` should not schedule rerenders
Historically, setting a tracked property triggered:

1. Invalidating the tracked property’s tag.
2. Scheduling a re-render.

However, in at least the case of a component’s `args` property,
scheduling a re-render is not desirable. That’s because `args` are set
by Glimmer during the render process, and a mutation never indicates
that the DOM may update.

There _may_, however, be derived, computed state that should be invalidated
by the changing `args`, which should happen before the component is updated.
In that case, the tracked property semantics of invalidating a tag are desired,
because it unlocks the ability to invalidate a component’s tracked computed
properties.

This commit changes the component’s `args` property to use a special
getter/setter instead of `@tracked`. This setter manually invalidates
the property’s tag but does not schedule a re-render.