Implement component `bounds` feature
With the recent Glimmer VM changes to adopt IndustryStandardCase, Glimmer.js components now relax the requirement of a template having a single root element. In addition to having multiple elements, this also means that a template may have top-level nodes such as CommentNodes or TextNodes.

This change renders the previous `element` API incoherent because there is no guarantee of a single "main" element for a template. Instead, we now set the `bounds` property of a component, which contains pointers to the first node and last node.

This is obviously less convenient for cases where you do just have a single conceptual main template. For example, if you have a template with a single root element, you would probably prefer to use `this.element` instead of `this.bounds.firstNode` or `this.bounds.lastNode`.

We could automatically determine the "single element" case and assign element automatically, but we were afraid that it would become a refactoring hazard. If you had a template with a single root element, wrote a bunch of code that used `this.element`, and then someone came along and added a comment to the beginning, all of that code would suddenly stop working.

Instead, you can now indicate which element is the conceptual "main" element by using `...attributes`, which also is the location where invocation-side attributes will get splatted into.
For example:

```hbs
{{!-- MyComponent.hbs --}}
<h1>Title</h1>
<div class="content" ...attributes>
</div>
```

In this example, `this.element` would be the `<div>` element. If you invoked this component with `<MyComponent id="component-id" />`, this `<div>` would also receive the `id` attribute.