Introduce @glimmer/compiler-delegates
This introduces some builtin compile package for compiler delegates that can be used in the bundle compiler. A compiler delegate is a class that encapsulates all of the logic for answering questions about the app code on behalf of the bundle compiler. For instance the Module Unification Compiler Delegate encapsulates all the lookup rules for a project using MU. The delegates are meant to be generic in the sense that they are not coupled to a specific build pipeline.