Introduce Application Delegates
As the Glimmer VM has continued to expand its platform of capabilties it has become clear that supporting a matrix of modes would be very difficult and confusing for users. Instead we should have sane defaults but provide the primitives to changing how a glimmer.js application behaves. Below is a matrix of capabilties that the VM currently supports:

| Mode | Client | Server |
|------|--------|--------|
|Rehydration| X | Requires rehydration serialization |
|ClientSide Rendering| X | |
|LazyCompilation| X | X |
|EagerCompilation| X | X |
|StringBuilding (SEO)| | X |

As you can see this matrix is pretty large. To deal with this, this PR introduces the idea of Application Delegates to augment how a glimmer.js application should work. This initial PR first seperates out the destict pieces around rendering.