didRender helper shouldn't poll
Previously, the didRender test helper was polling every 10ms to see whether a render completed. For hopefully obvious reasons, this leads to the potential for race conditions if the browser starts throttling setInterval or multiple renders complete in under 10ms. We should strive to avoid non-determinism in our test suite.

This commit changes didRender to register for notifications from the application. Notifiers are invoked after every render and re-render. As a bonus, we now propagate errors through the helper, whereas before they would be swallowed and the didRender promise would simply never resolve.