fix type error in application/src/environment.ts
`item` param in `keyFor` method is an `Opaque type` that can be either
an `object`, `void`, `null` or `undefined`, therefore has no index
signature, so ts considere an implicit `any` when used like
`item[keypath]`.

a solution is to force cast `item`as `any` then return whatever we got
as `String` to ensure the type is correct.