Make "main" actually work
Prior to this commit there were 2 rather large issues.

1. The code generation was not in the format that the VM was expecting
2. I merged the "main" code in Glimmer VM and it wasn't integrated at all

Fixing the code generation was pretty straight forward however implimenting
the main is a bit tricky as it requires to know where the main template is.
The 'main' is just a handle to an operation that does an invoke of a library
supplied mainLayout. This makes this a bit tricky as you need to actually load
the VM stack with the correct obejcts before the VM actually executes. This means
we need the MU specifier somewhere in the data payload so we can use it to lookup
it's handle and symbols. We likely need to think about this contract a bit.