Introduce convention of where the main template is
This introduces the concept of a well-known main template. A template that
is added to the Bundle Compiler with the locator of:

```
{ module: '@glimmer/application', name: 'mainTemplate' }
```

This will become special cased specifier known as "mainTemplate" in both the data segment.
With the introduction of the main opcode in the VM we need to have serialized info
about layout that the main opcode is invoking. Specifically we need the
ability to lookup `ProgramSymbolTable` and `handle` for the main layout
at runtime. Without having this convention we have no idea what to actually invoke as
the main layout.