Include handle in bytecode data segment metadata
Currently, we ship a JSON data structure with bytecode templates called `meta` that is keyed on specifier string and contains the component's VM handle and symbol table. This data structure supports dynamic lookup of components via the `{{component}}` helper, even though most component invocations are precompiled in bytecode mode.

However, we currently don't serialize each component's handle, which is an application-specific identifier. Handles are used to allow opcodes to refer to "live" JavaScript objects, like component classes, at runtime. While handles are built into the compiled bytecode, they are not accessible to dynamic component invocation. This means that Glimmer.js apps have to ship both the module resolution map as well as the template metadata just to get an app to boot, which is wasteful.

This commit adds each component's handle to its metadata entry in the compiled data segment. It also updates the BytecodeResolver to use the handle and external module table, so resolution happens identically to the static invocation path. In addition to this resolution being faster, it also means that Glimmer.js apps can boot without a resolution map and use the bytecode metadata for everything.