Support incremental rebuilds in app compiler
This commit solves one of the hardest problems in computer science. The BundleCompiler from Glimmer VM is stateful and is only intended to be used for one-shot compilation. Unfortunately the current implementation of the app compiler caches the bundle compiler across rebuilds, meaning that you have to start and stop the Broccoli server to see changes. This commit changes the implementation to create a fresh compiler/delegate every rebuild.