Use WeakMap to track tracked property meta
Tracked properties rely on storing unique tags for each property on an object. We store these and other metadata in instances of a class called Meta, which are created lazily the first time a tracked property is requested for a property on an object.

Previously, we used a symbol to store the object's Meta on itself. The symbol prevented inadvertent access or naming collisions. However, setting a property, even via a symbol, throws an error if the object in question is frozen via `Object.freeze`. It is a common pattern for libraries to use `Object.freeze` on objects that should be read-only, such as Apollo (#632). Simply reading a property from an object should not cause an exception.

To mitigate this, this commit changes the implementation to use a WeakMap to map objects to their corresponding Meta instance. The new implementation is intended to be semantically identical to the previous implementation, but with compatibility with frozen objects.