Run tests in strict mode
Previously, tests were not running in strict mode, because we were emitting them as a JavaScript module (which implies strict mode) but not consuming them as such, so they defaulted to sloppy mode.

In practice, this meant that several tests that verified compatibility with frozen objects were inappropriately passing, because mutation of frozen objects in sloppy mode is a no-op rather than an error.

This commit changes the Rollup configuration for tests to use `iife` as the target format rather than `es`, which produces the appropriate `"use strict";` pragma.