Remove scheduleRerender monkeypatch in testing
At some point we started monkeypatcihng scheduleRerender in the tests to detect when a re-render had finished. However, this was inappropriate because it predictably caused a divergence between the test implementation and the real implementation. Additionally, we already have a system for observing render completion via the `didRender` function.

This commit reverts the scheduleRerender monkeypatch and updates a handful of tests relying on it to use the existing didRender helper instead.