[FEAT] Implement Glimmer Components RFC
This PR turns the `@glimmer/component` package into an Ember addon,
allowing it to work in both Ember.js and Glimmer.js applications. This
works by keeping the `GlimmerComponent` class definition in a separate
file that then gets imported by the Glimmer.js implementation in `src`,
and is included in the Ember.js implementation via `addon/index.js`.