Re-enable lifecycle hooks for Glimmer.js environments
This commit restores the `didInsertElement` and `didUpdate` lifecycle hooks and `element`/`bounds` properties that were removed during the work to enable @glimmer/component to work as an Ember addon.

Our intent is for developers to use an element modifier-based API to access the DOM, rather than component hooks and properties. This is supported in the Ember environment today. However, the required element modifier manager API is not yet implemented in Glimmer.js.

Not having this capability blocks many Glimmer.js apps from upgrading, so this commit restores that functionality and allows those apps to upgrade to 0.14. Once we implement element modifier manager support in Glimmer.js, we can re-remove these APIs and align the Glimmer/Ember environments.