[BUGFIX] Fixes args tracking in Glimmer component
`@tracked` is not yet avalibale for general usage in Ember apps,
so we were relying on using the interop with `get` and
`set` to track args. However, we never actually used `get`,
so it wasn't picking it up. This ensures that all gets/sets of
the field go through the proper codepaths, and will be autotracked.