[BUGFIX] Updates the assignment of args to reflect the arg proxy
The refactor to use a proxy for tracking individual updates to args
requires us to make some changes here in how we assign arguments. Also
adds the `updateHook` option to capabilities. This solution should be
compatible with Ember pre and post enabling the tracked canary flag,
but can be made better with ember-compatibility-helpers once we have
a version of Ember to target.