Remove need for @tracked on getters
This change aligns the Glimmer.js @tracked implementation with Ember, where further improvements to the design have been made.

In particular, values consumed through templates (as represented by References) now participate in the autotracking system. This removes the need for annotating getters with the @tracked decorator. Instead, getters become "tracked" automatically as a by-product of accessing tracked properties during their execution.

As of this change, applying @tracked to a getter will produce an error message with instructions to remove the decorator.