[REFACTOR] Removes most of the remaining custom reference code.
Removes most of the remaining custom reference code, with the exception
of helper references (which will be factored out soon).

Also removes the tracking/rendering assertion for the time being,
pending discussion on whether or not the assertion continues to make
sense. If the discussion concludes it does, we can add it back before
the v2 release, with more thorough integration testing that actually
tests the expected behavior, rather than unit tests.