[FEATURE] Adds Owner
Adds a `owner` argument that can be passed to `renderComponent`. This
value is provided to every manager in that render tree via the manager
factory:

```js
setComponentManager((owner) => new CustomManager(owner), {});
```

Conceptually, the owner here is similar to the owner in Ember.js.
Glimmer.js creates a new manager for each unique owner that it receives,
and managers can use the `getOwner` and `setOwner` methods to assign it
to passed objects. The owner is an opaque value, so it's up to the
embedding environment to decide what its API is. If no owner is passed,
a default owner is used, and throws an error if the user attempts to use
it for anything.