[CLEANUP] Updates ComponentFactory to ComponentDefinition
As I began the work on adding `ModifierManager` and `HelperManager`, it
became more clear that "factory" wasn't really the right term for the
values provided to `setComponentManager`, etc. We aren't assigning a
factory, we're assigning a _definition_ that could be a factory.

The only reason this would be an issue is it somewhat conflicts with the
terminology used internally by the VM. This PR makes the distinction
more clear, and renames "factory" to "definition" for the external
types in general.