[BUGFIX] Ensures Strict Mode
Ensures that non-strict-mode lookups will fail in development builds,
preventing sloppy mode code. This ensures that the API for templates
will be stable as we continue developing strict mode within the VM.

The way it does this currently is to parse the template in DEBUG builds
and compare the upvars within it to the `scope` provided to the
template. If a value isn't found in scope, and is not a built-in, then
it throws.