[BUGFIX] Ensure Owner is generic for setManager APIs
Currently the `owner` parameter for these APIs is typed as `object`,
because that's the only minimum requirement (since we associate the
instance to the owner via a WeakMap). This is fine, unless the user
expects a specific type of owner class (e.g. they are specifying their
own).

This PR makes the `owner` parameter generic all the way through the
interfaces, so users can provide their own interface as long as it
extends `object` (and satisfies the WeakMap key requirements).