[BUGFIX] Binds EnvironmentDelegate.protocolForUrl
We made a decision upstream to pluck the delegate functions and move
them around so that they could be looked up more quickly when used. In
most cases this is fine, because the delegate functions are pure and
non-stateful. The `protocolForUrl` function is not, but this was known
at the time, and the plan was to bind it like in this PR. It was an
oversight in the upgrade process.

Tests added to ensure we remember if future changes occur.