Introduce Component `Signature` (RFC 0748)
Adds backwards-compatible support for component `Signature`s per [Ember
RFC #0748][rfc]. The public API of `Component` is now *more* permissive
than it was previously, because it changes the type parameter to be an
unconstrained generic `<S>` (for "signature") which can accept *either*
the existing `Args` types *or* a new `Signature` which includes `Args`
and adds `Blocks` and `Element`.

[rfc]: https://github.com/emberjs/rfcs/pull/748

The `Args` part of the new signature work exactly like the old
args-only type did. The `Blocks` and `Element` parts of a signature are
inert from the perspective of TypeScript users who are not yet using
[Glint][glint], but Glint users will benefit directly once Glint
releases an update which can requires a version of `@glimmer/component`
incorporating this change.

[glint]: https://github.com/typed-ember/glint

Since this change is backwards compatible, we can deprecate the
non-`Signature` form at a later time when we are ready for a 2.0
release.

To validate these changes, with the relatively complicated type
machinery they require under the hood, this also introduces the
`expect-type` type testing library, rewrites the existing type tests
using it, and introduces new type tests for all supported forms of the
`Signature`.

(cherry picked from commit 967d028de585fd0175c72af1ed96795b0018895e)