Convert entirely to expect-type for type tests
dtslint is built on the deprecated tslint, and is finicky as can be, and
it proved impossible to get the existing dtslint setup actually working
with the places we need to use `@ts-ignore` pragmas (to account for TS
improvements landed since 3.5)

Preferably, we should add checking against a range of supported TS
versions to go with this on CI in the future. Or just actually release
v2.0 and get away from this entirely!