Update import for private type location for build results
NOTE: this *will not* type-check until you have done a build, because
it is the result of emitting types to the production build location.
However, this matches the actual import location to which this type
would be emitted. Since this is an internal-only type whose presence
consumers should not rely on and which they should not use in any way,
this is "safe" from a public API POV.